requirements.txt
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
cycler==0.11.0
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.28.5
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.0
packaging==21.3
pandas==1.3.5
paramiko==2.9.2
Pillow==9.0.0
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.29
sshtunnel==0.4.0
toml==0.10.2
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.1
